+++ docs/release-notes.rst
@@ -14,6 +14,7 @@ Fixed bugs
 ----------
 
 - There was no error when trying to execute a file that is not a JS file (issue #67)
+- Errors appearing during the execution of the callback of ``webpage.open()`` are now displayed
 
 
 version 0.8.1